Be a little more careful with the shift computations in QT2QTN and makeTSQuerySign.  The first of these is a live bug, on some platforms, as per bug #5590 from John Regehr.  However the consequences seem limited because of the relatively narrow scope of use of QTNode.sign.  The shift in makeTSQuerySign is actually safe because TSQS_SIGLEN is unsigned, but it seems like a good idea to insert an explicit cast rather than depend on that.

+++ src/backend/utils/adt/tsquery_op.c
@@ -209,7 +209,7 @@ makeTSQuerySign(TSQuery a)
 	for (i = 0; i < a->size; i++)
 	{
 		if (ptr->type == QI_VAL)
-			sign |= ((TSQuerySign) 1) << (ptr->qoperand.valcrc % TSQS_SIGLEN);
+			sign |= ((TSQuerySign) 1) << (((unsigned int) ptr->qoperand.valcrc) % TSQS_SIGLEN);
 		ptr++;
 	}
